fix: fix jest test error by fixing import path (TNA-86)
